# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enterprise Headless CMS API",
    "description": "Enterprise headless CMS backend for multi-site content modeling, editorial workflows, media, localization, delivery APIs, webhooks, API keys, navigation and analytics.",
    "contact": {
      "name": "CMS Platform Team"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/content-types": {
      "post": {
        "tags": [
          "Content Types"
        ],
        "summary": "Create Content Type",
        "operationId": "create_content_type_api_v1_content_types_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentTypeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentTypeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Content Types"
        ],
        "summary": "List Content Types",
        "operationId": "list_content_types_api_v1_content_types_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContentTypeOut"
                  },
                  "title": "Response List Content Types Api V1 Content Types Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content-types/{content_type_id}": {
      "get": {
        "tags": [
          "Content Types"
        ],
        "summary": "Get Content Type",
        "operationId": "get_content_type_api_v1_content_types__content_type_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "content_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Type Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentTypeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Content Types"
        ],
        "summary": "Update Content Type",
        "operationId": "update_content_type_api_v1_content_types__content_type_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "content_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Type Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentTypeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentTypeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Content Types"
        ],
        "summary": "Archive Content Type",
        "operationId": "archive_content_type_api_v1_content_types__content_type_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "content_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Type Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content-types/{content_type_id}/publish": {
      "post": {
        "tags": [
          "Content Types"
        ],
        "summary": "Publish Content Type",
        "operationId": "publish_content_type_api_v1_content_types__content_type_id__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "content_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Type Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentTypeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content-types/{content_type_id}/duplicate": {
      "post": {
        "tags": [
          "Content Types"
        ],
        "summary": "Duplicate Content Type",
        "operationId": "duplicate_content_type_api_v1_content_types__content_type_id__duplicate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "content_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Type Id"
            }
          },
          {
            "name": "new_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "New Key"
            }
          },
          {
            "name": "new_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "New Name"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentTypeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentTypeCreate": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "key": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_-]{1,78}$",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "schema_definition": {
            "additionalProperties": true,
            "type": "object",
            "title": "Schema Definition"
          },
          "display_field": {
            "type": "string",
            "title": "Display Field",
            "default": "title"
          },
          "is_singleton": {
            "type": "boolean",
            "title": "Is Singleton",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "key",
          "name",
          "schema_definition"
        ],
        "title": "ContentTypeCreate"
      },
      "ContentTypeOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "schema_definition": {
            "additionalProperties": true,
            "type": "object",
            "title": "Schema Definition"
          },
          "display_field": {
            "type": "string",
            "title": "Display Field"
          },
          "is_singleton": {
            "type": "boolean",
            "title": "Is Singleton"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "site_id",
          "key",
          "name",
          "description",
          "schema_definition",
          "display_field",
          "is_singleton",
          "is_published",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "ContentTypeOut"
      },
      "ContentTypeUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "schema_definition": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Definition"
          },
          "display_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Field"
          },
          "is_singleton": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Singleton"
          }
        },
        "type": "object",
        "title": "ContentTypeUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Message"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}